server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
fastapi==0.123.5
filelock==3.20.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6
numpy==2.3.5
openai==2.8.1
packaging==25.0
PyAudio==0.2.14
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
setuptools==80.9.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
vad==1.0.2
